+++ frontend/src/components/modals/index.js
@@ -8,5 +8,6 @@
 
 export { default as BaseModal } from './BaseModal.vue';
 export { default as ConfirmModal } from './ConfirmModal.vue';
+export { default as StationFormModal } from './StationFormModal.vue';
 export { default as PlatformFormModal } from './PlatformFormModal.vue';
 export { default as InstrumentFormModal } from './InstrumentFormModal.vue';